# src/go/internal/models/injector/clear_sidecar_request.go
package injectormodels

type ClearSidecarPayload struct {
	Namespace            string `json:"Namespace" binding:"required"`
	DeploymentName       string `json:"DeploymentName" binding:"required"`
	SidecarContainerName string `json:"SidecarContainerName" binding:"required"`
}